requirements.txt
asgiref==3.6.0
Django==4.2
django-ckeditor==6.5.1
django-filter==23.1
django-js-asset==2.0.0
djangorestframework==3.14.0
Markdown==3.4.3
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
